backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic==1.10.14
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
alembic==1.13.1
httpx==0.26.0
pytest==7.4.4
